move tests out of production
